bmi view

+++ src/main/java/com/github/anastr/speedviewlib/util/OnPrintTickLabel.java
@@ -10,5 +10,5 @@ public interface OnPrintTickLabel {
      * @param tick speed value at the tick.
      * @return label to draw.
      */
-    String getTickLabel(int tickPosition, int tick);
+    String getTickLabel(int tickPosition, float tick);
 }
